Minor clear: - Cleaned main.cpp - Tested compilation up to c++26 (fixed usage of size) - .gitattributes to fix stats - Minor Readme adjustments
