+++ src/core/cft.hpp
@@ -94,14 +94,14 @@ constexpr real_t operator""_F(long double f) {
 // container as cidx_t. It also allows to avoid narrowing conversion warnings.
 template <typename Cont>
 inline cidx_t csize(Cont const& cont) {
-    return as_cidx(size(cont));
+    return as_cidx(cft::size(cont));
 }
 
 // Since ridx_t could be any integer type, rsize provide a (debug) checked way to get the size of a
 // container as ridx_t. It also allows to avoid narrowing conversion warnings.
 template <typename Cont>
 inline ridx_t rsize(Cont const& cont) {
-    return as_ridx(size(cont));
+    return as_ridx(cft::size(cont));
 }
 
 struct CidxAndCost {
